Maurer Rose in p5.js
Maurer rose was introduced by Peter M. Maurer in his article titled A Rose is a Rose...

Let r = sin(nθ) be a rose in the polar coordinate system, where n is a positive integer. 
  The rose has n petals if n is odd, and 2n petals if n is even.

   We then take 361 points on the rose:

  (sin(nk), k) (k = 0, d, 2d, 3d, ..., 360d),
  where d is a positive integer and the angles are in degrees, not 
  radians.

A Maurer rose of the rose r = sin(nθ) consists of 360 lines successively connecting the 361 points on a polar plane.